+++ composables/fasttrak-api/index.ts
@@ -1,3 +1,2 @@
 export * from './schemas'
 export * from './utils'
-export * from './types'
