+++ composables/fasttrak-api/schemas/index.ts
@@ -2,8 +2,8 @@ export * from './AgingReportSchema'
 export * from './AuthResponseSchema'
 export * from './CustomerSchema'
 export * from './DriverSchema'
+export * from './EmployeePaySchema'
 export * from './EmployeeSchema'
 export * from './PricingPlanSchema'
 export * from './ReservationSchema'
-export * from './SingleReservation'
 export * from './VehicleTypeSchema'
